+++ src/Traits/QueriesFlaggedEnums.php
@@ -2,6 +2,7 @@
 
 namespace BenSampo\Enum\Traits;
 
+use BenSampo\Enum\FlaggedEnum;
 use Illuminate\Database\Eloquent\Builder;
 
 trait QueriesFlaggedEnums
@@ -10,7 +11,7 @@ trait QueriesFlaggedEnums
      * @param  \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model> $query
      * @return \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model>
      */
-    public function scopeHasFlag(Builder $query, string $column, int $flag): Builder
+    public function scopeHasFlag(Builder $query, string $column, int|FlaggedEnum $flag): Builder
     {
         return $query->whereRaw("$column & ? > 0", [$flag]);
     }
@@ -19,32 +20,47 @@ trait QueriesFlaggedEnums
      * @param  \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model> $query
      * @return \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model>
      */
-    public function scopeNotHasFlag(Builder $query, string $column, int $flag): Builder
+    public function scopeNotHasFlag(Builder $query, string $column, int|FlaggedEnum $flag): Builder
     {
         return $query->whereRaw("not $column & ? > 0", [$flag]);
     }
 
     /**
      * @param  \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model> $query
-     * @param  array<int> $flags
+     * @param  array<int|FlaggedEnum> $flags
      * @return \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model>
      */
     public function scopeHasAllFlags(Builder $query, string $column, array $flags): Builder
     {
-        $mask = array_sum($flags);
+        $mask = $this->flagsSum($flags);
 
         return $query->whereRaw("$column & ? = ?", [$mask, $mask]);
     }
 
     /**
      * @param  \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model> $query
-     * @param  array<int> $flags
+     * @param  array<int|FlaggedEnum> $flags
      * @return \Illuminate\Database\Eloquent\Builder<\Illuminate\Database\Eloquent\Model>
      */
     public function scopeHasAnyFlags(Builder $query, string $column, array $flags): Builder
     {
-        $mask = array_sum($flags);
+        $mask = $this->flagsSum($flags);
 
         return $query->whereRaw("$column & ? > 0", [$mask]);
+    }
+
+    /**
+     * @param  array<int|FlaggedEnum> $flags
+     */
+    protected function flagsSum(array $flags): int
+    {
+        return array_reduce(
+            $flags,
+            static fn(int $carry, int|FlaggedEnum $flag): int => $carry
+                + ($flag instanceof FlaggedEnum
+                    ? $flag->value
+                    : $flag),
+            0
+        );
     }
 }